create login panel protected route to redirect when user is loged in and moved protectedRoute to lib
