+++ client/src/lib/ProtectedRoute.tsx
@@ -1,4 +1,4 @@
-import { isAuthenticating, selectCurrentUser } from "@/store/authSlice";
+import { selectIsAuthenticating, selectCurrentUser } from "@/store/authSlice";
 import { useSelector } from "react-redux";
 import { Navigate } from "react-router-dom";
 
@@ -8,7 +8,7 @@ export default function ProtectedRoute({
   children: React.ReactNode;
 }) {
   const user = useSelector(selectCurrentUser);
-  const authenticating = useSelector(isAuthenticating);
+  const authenticating = useSelector(selectIsAuthenticating);
 
   if (!user && !authenticating) return <Navigate to="/auth" replace />;
   return children;
